Updated qs compatability, new strings for it in config.lua.
This is setup for new qs-smartphone to search the entire database for this. It was the only way possible at this time to do it. OldQsPhone is now compatible, I didnt know till recently it saves to playerdata like the standard phone.  --- let me know if you run into issues.

Thank you Quasar for sending me the phone so I could add compatibility easier and test.